requirements.txt
# Web Framework
flask==3.0.0

# AI és Embedding Eszközök
sentence-transformers==2.6.1
chromadb==0.4.24
langchain==0.1.0
langchain-community==0.0.29
langchain-anthropic==0.1.1

# Egyéb segédkönyvtárak
python-dotenv==1.0.0
markdown==3.5.2

# Fejlesztői eszközök
pytest==7.4.4
logging==0.5.1.2